as_ascii_table

This function takes a data.frame and attempts to convert it into a simple ascii format suitable for printing to the screen It is assumed all variable values have a as.character() method in order to cast them to character.

Usage

as_ascii_table(dat, line_prefix = "  ")

Arguments

dat

Input dataset to convert into a ascii table

line_prefix

Symbols to prefix in front of every line of the table