identify_differences

Compares each column within 2 datasets to identify any values which they mismatch on.

identify_differences(BASE, COMP, KEYS, exclude_cols,
  tolerance = sqrt(.Machine$double.eps), scale = NULL)

Arguments

BASE
Base dataset for comparison (data.frame)

COMP
Comparator dataset to compare base against (data.frame)

KEYS
List of variables that define a unique row within the datasets (strings)

exclude_cols
Columns to exclude from comparison

tolerance
Level of tolerance for numeric differences between two variables

scale
Scale that tolerance should be set on. If NULL assume absolute